demo_mod operational; pickling moved.
